Extraction_System_new/requirements.txt
PyMuPDF==1.24.10
PyYAML==6.0.2
regex==2024.7.24
rapidfuzz==3.9.6
pandas==2.2.2
numpy==1.26.4
pdfplumber==0.9.0